+++ Purple Robot Manager/src/edu/northwestern/cbits/purple_robot_manager/probes/builtin/RobotHealthProbe.java
@@ -406,7 +406,7 @@ public class RobotHealthProbe extends Probe
 		formatted.putLong(context.getString(R.string.robot_runtime_label), bundle.getLong(RobotHealthProbe.ACTIVE_RUNTIME, 0));
 		formatted.putFloat(context.getString(R.string.robot_cpu_load_label), bundle.getFloat(RobotHealthProbe.CPU_USAGE, 0));
 		formatted.putLong(context.getString(R.string.robot_time_offset_label), bundle.getLong(RobotHealthProbe.TIME_OFFSET_MS, 0));
-		formatted.putInt(context.getString(R.string.robot_pending_count_label), bundle.getInt(RobotHealthProbe.PENDING_COUNT, 0));
+		formatted.putInt(context.getString(R.string.robot_pending_count_label), (int) bundle.getDouble(RobotHealthProbe.PENDING_COUNT, 0));
 		formatted.putLong(context.getString(R.string.robot_pending_size_label), bundle.getLong(RobotHealthProbe.PENDING_SIZE, 0));
 		formatted.putLong(context.getString(R.string.robot_clear_time_label), bundle.getLong(RobotHealthProbe.CLEAR_TIME, 0));
 
@@ -417,17 +417,17 @@ public class RobotHealthProbe extends Probe
 	
 	public String summarizeValue(Context context, Bundle bundle)
 	{
-		int count = bundle.getInt(RobotHealthProbe.PENDING_COUNT);
-		double size = 0.0 + bundle.getLong(RobotHealthProbe.PENDING_SIZE);
+		int count = (int) bundle.getDouble(RobotHealthProbe.PENDING_COUNT);
+		double size = 0.0 + (int) bundle.getDouble(RobotHealthProbe.PENDING_SIZE);
 
-		long clear = bundle.getLong(RobotHealthProbe.CLEAR_TIME);
+		long clear = (int) bundle.getDouble(RobotHealthProbe.CLEAR_TIME);
 
 		if (clear < 0)
 			clear = 0;
 
 		size = size / (1024 * 1024);
 
-		float cpu = bundle.getFloat(RobotHealthProbe.CPU_USAGE);
+		double cpu = bundle.getDouble(RobotHealthProbe.CPU_USAGE);
 
 		return String.format(context.getResources().getString(R.string.summary_robot_probe), cpu, count, size, clear);
 	}
